Detalles de cuenta terminado

+++ Proyect_app/app/src/main/java/com/example/proyecto_final_apps/data/repository/OperationRepositoryImp.kt
@@ -6,8 +6,10 @@ import com.example.proyecto_final_apps.data.local.Database
 import com.example.proyecto_final_apps.data.local.MyDataStore
 import com.example.proyecto_final_apps.data.local.entity.OperationModel
 import com.example.proyecto_final_apps.data.remote.API
-import com.example.proyecto_final_apps.data.remote.dto.operationResponse.toOperationModel
+import com.example.proyecto_final_apps.data.remote.dto.getOperationsResponse.toOperationModel
 import com.example.proyecto_final_apps.helpers.DateParse
+import com.example.proyecto_final_apps.helpers.Internet
+import java.util.*
 import javax.inject.Inject
 
 class OperationRepositoryImp @Inject constructor(
@@ -18,40 +20,40 @@ class OperationRepositoryImp @Inject constructor(
 
     override suspend fun getOperations(forceUpdate: Boolean): Resource<List<OperationModel>> {
 
-        try{
+        try {
 
-        val operationsStoredNumber = database.operationDao().getOperationsStoredNumber()
-        if (operationsStoredNumber > 0 && !forceUpdate) {
+            val operationsStoredNumber = database.operationDao().getOperationsStoredNumber()
+            if (operationsStoredNumber > 0 && !(forceUpdate && Internet.checkForInternet(context))) {
 
-            //local data
-            val operationsList = database.operationDao().getAllOperations()
-            return Resource.Success(operationsList)
-        } else {
+                //local data
+                val operationsList = database.operationDao().getAllOperations()
+                return Resource.Success(operationsList)
+            } else {
 
-            val ds = MyDataStore(context)
-            val token = ds.getValueFromKey("token") ?: return Resource.Error("No token")
+                val ds = MyDataStore(context)
+                val token = ds.getValueFromKey("token") ?: return Resource.Error("No token")
 
-            //From api
-            val result = api.getAllOperations(token)
+                //From api
+                val result = api.getAllOperations(token)
 
-            if (result.isSuccessful) {
-                val operationsList = result.body()?.operations?.map { it.toOperationModel() }
-                if (operationsList != null && operationsList.isNotEmpty()){
+                if (result.isSuccessful) {
+                    val operationsList = result.body()?.operations?.map { it.toOperationModel() }
+                    if (operationsList != null && operationsList.isNotEmpty()) {
 
-                    //store in db
-                    database.operationDao().deleteAllOperations()
-                    database.operationDao().insertMany(operationsList)
+                        //store in db
+                        database.operationDao().deleteAllOperations()
+                        database.operationDao().insertMany(operationsList)
 
-                    return Resource.Success(operationsList)
-                }else
-                    return Resource.Error("No hay operaciones por mostrar.")
+                        return Resource.Success(operationsList)
+                    } else
+                        return Resource.Error("No hay operaciones por mostrar.")
 
-            } else
-                println(result.message())
+                } else
+                    println(result.message())
 
-        }
+            }
 
-        }catch(ex:Exception){
+        } catch (ex: Exception) {
 
         }
 
@@ -79,17 +81,17 @@ class OperationRepositoryImp @Inject constructor(
         val balanceResult = getGeneralBalance()
         val operationsList = database.operationDao().getAllOperations()
 
-        if( balanceResult is Resource.Success && operationsList.isNotEmpty()){
+        if (balanceResult is Resource.Success && operationsList.isNotEmpty()) {
             val currentBalance = balanceResult.data
             var lastBalance = 0.0
 
 
-            operationsList.forEach{ op ->
+            operationsList.forEach { op ->
 
                 val operationDate = DateParse.formatDate(op.date)
                 val firstDayMonthDate = DateParse.getFirstDayOfMonthDate()
 
-                if(firstDayMonthDate > operationDate){
+                if (firstDayMonthDate > operationDate) {
                     if (op.active) lastBalance += op.amount
                     else lastBalance -= op.amount
                 }
@@ -100,5 +102,64 @@ class OperationRepositoryImp @Inject constructor(
         }
 
         return Resource.Success(0.00)
+    }
+
+    private fun filterOperationByDate(
+        operations: List<OperationModel>,
+        startDate: Date?,
+        endDate: Date?
+    ): List<OperationModel> {
+
+        return operations.filter {
+            val operationDate = DateParse.formatDate(it.date)
+            if (startDate != null && endDate != null) {
+                operationDate >= startDate && operationDate <= endDate
+            } else if (startDate != null && endDate == null)
+                operationDate >= startDate
+            else if (endDate != null)
+                operationDate <= endDate
+            else true
+        }
+
+    }
+
+    override suspend fun getAccountOperations(
+        localAccountId: Int,
+        forceUpdate: Boolean,
+        startDate: Date?,
+        endDate: Date?
+    ): Resource<List<OperationModel>> {
+
+        var storedOperations = database.operationDao().getAccountOperations(localAccountId)
+
+        if (storedOperations.isNotEmpty() && !(forceUpdate && Internet.checkForInternet(context))) {
+
+            //filtrar y retornar operaciones de la bd
+            val filteredOperations = filterOperationByDate(storedOperations, startDate, endDate)
+            return if (filteredOperations.isNotEmpty())
+                Resource.Success(filteredOperations)
+            else Resource.Error("No se encontraron operaciones dentro del rango de fechas.")
+        } else {
+            //download operations from api
+            when (val result = getOperations(true)) {
+                is Resource.Success -> {
+
+                    //filtrar y retornar operaciones de la bd
+                    storedOperations = database.operationDao().getAccountOperations(localAccountId)
+                    return if (storedOperations.isNotEmpty()) {
+                        val filteredOperations =
+                            filterOperationByDate(storedOperations, startDate, endDate)
+                        if (filteredOperations.isNotEmpty())
+                            Resource.Success(filteredOperations)
+                        else Resource.Error("No se encontraron operaciones dentro del rango de fechas.")
+
+                        //No se encontró ninguna operacion de la cuenta
+                    } else Resource.Error("No operations founded.")
+
+                }
+                else -> return result //Estado de error de getOperations
+            }
+        }
+
     }
 }
